+++ PuffyDonut/src/main/java/repositories/interfaces/OrmRepository.java
@@ -7,7 +7,7 @@ import java.util.Optional;
 
 public interface OrmRepository<T> {
     List<T> findAll();
-    T findById(Long id);
+    T findById(Integer id);
     void save(T entity);
 
 }
