+++ PuffyDonut/src/main/java/services/interfaces/DeliveryService.java
@@ -5,5 +5,5 @@ import models.Delivery;
 public interface DeliveryService {
     void addDelivery(Delivery delivery);
 
-    Delivery getDeliveryById(Long id);
+    Delivery getDeliveryById(Integer id);
 }
